+++ freediscovery/cluster/__init__.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
 
-from .base import (select_top_words, ClusterLabels,
-                   _ClusteringWrapper)
+from .base import (select_top_words, ClusterLabels)
 from .utils import centroid_similarity
 from .optimal_sampling import compute_optimal_sampling
